Header file untangling (part 3).
