+++ src/HealthGPS/repository.h
@@ -7,6 +7,7 @@
 #include "lms_definition.h"
 #include "modelinput.h"
 #include "riskfactor_adjustment_types.h"
+#include "riskfactor_model.h"
 #include <functional>
 #include <memory>
 #include <mutex>
